move 25 min time window change in time calc to add features
